PIOutputMapping

An object that combines simulation output with observed data. The parameter identification minimizes the distance between the simulation output and the observed data.

Super class

ospsuite.utils::Printable -> PIOutputMapping

Active bindings

observedXYData

Named list with the XYData that will be compared with simulation results. Names are the labels of the xySeries objects

quantity

Simulation quantity which results are to be compared to observed data. Read-only.

transformResultsFunction

Function that will be applied to simulated results. Allows to manipulate simulated values before calculating the residuals. The function should manipulate numeric vectors 'xVals' and 'yVals' (that being the simulated time- and observation values) that will be than assigned to the x- and y-values of the simulated result. The function must return a named list with key 'xVals' and 'yVals'.

Methods

Public methods

PIOutputMapping$new()

PIOutputMapping$addObservedData()

PIOutputMapping$removeXYSeries()

PIOutputMapping$setXFactors()

PIOutputMapping$setYFactors()

PIOutputMapping$setXOffset()

PIOutputMapping$print()

PIOutputMapping$clone()

Inherited methods

Method new()

Initialize a new instance of the class

Usage

PIOutputMapping$new(quantity)

Arguments

quantity

An object of the type Quantity

Returns

A new PIOutputMapping object. Add observed data as XYData object(s).

Method addObservedData()

Usage

PIOutputMapping$addObservedData(XYData)

Arguments

XYData

Object or a list of objects of the type XYData. Each data set must be of the same dimension as the simulation quantity of the mapping.

Details

If an observed data object with the same label already exists, it will be overwritten.

Method removeXYSeries()

Remove the observed data.

Usage

PIOutputMapping$removeXYSeries(label)

Arguments

label

label of the x-y values series to be removed

Method setXFactors()

Set the X-factors of x-y values by labels.

Usage

PIOutputMapping$setXFactors(labels, xFactors)

Arguments

labels

A list of label of XYData

xFactors

Numeric values that will be multiplied by the x-values

Method setYFactors()

Set the y-factors of x-y values by labels.

Usage

PIOutputMapping$setYFactors(labels, yFactors)

Arguments

labels

A list of label of XYData

yFactors

Numeric values that will be multiplied by the y-values

Method setXOffset()

Set the X-offset of x-y values by labels.

Usage

PIOutputMapping$setXOffset(labels, xOffset)

Arguments

labels

A list of label of XYData

xOffset

Numeric values that will be added to the x-values

Method print()

Print the object to the console

Usage

PIOutputMapping$print(...)

Arguments

...

Rest arguments.

Method clone()

The objects of this class are cloneable with this method.

Usage

PIOutputMapping$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.