PIOutputMapping

An object that links together quantities from the simulation to observed data. This object is passed to the ParameterIdentification objects

Super class

ospsuite.utils::Printable -> PIOutputMapping

Active bindings

observedDataSets

Named list of DataSet objects that will be compared to simulation results.

dataTransformations

a named list of factors and offsets

quantity

Simulation quantities which values are matched to the observed data

scaling

Linear (default) or logarithmic scaling for this output mapping

transformResultsFunction

Function that will be applied to simulated results. Allows to manipulate simulated values before calculating the residuals. The function should manipulate numeric vectors 'xVals' and 'yVals' (that being the simulated time- and observation values) that will be than assigned to the x- and y-values of the simulated result. The function must return a named list with key 'xVals' and 'yVals'.

Methods

Public methods

PIOutputMapping$new()

PIOutputMapping$addObservedDataSets()

PIOutputMapping$removeObservedDataSet()

PIOutputMapping$setDataTransformations()

PIOutputMapping$print()

PIOutputMapping$clone()

Method new()

Initialize a new instance of the class

Usage

PIOutputMapping$new(quantity)

Arguments

quantity

An object of the type Quantity

Returns

A new PIOutputMapping object. Add observed data as DataSet objects

Method addObservedDataSets()

Usage

PIOutputMapping$addObservedDataSets(data)

Arguments

data

Object or a list of objects of the type DataSet. Each data set must be of the same dimension as the simulation quantity of the mapping.

Details

If an observed data object with the same label already exists, it will be overwritten.

Method removeObservedDataSet()

Remove the observed data.

Usage

PIOutputMapping$removeObservedDataSet(label)

Arguments

label

label of the x-y values series to be removed

Method setDataTransformations()

Set the data transformations

Usage

PIOutputMapping$setDataTransformations(
  labels = NULL,
  xOffsets = 0,
  yOffsets = 0,
  xFactors = 1,
  yFactors = 1
)

Arguments

labels

A list of labels, each corresponding to one of the datasets. If labels are not specified, data transformations are set across all datasets

xOffsets

A numeric list or a value of X-offsets

yOffsets

A numeric list or a value of Y-offsets

xFactors

A numeric list or a value of X-factors

yFactors

A numeric list or a value of Y-factors

Method print()

Print the object to the console

Usage

PIOutputMapping$print(...)

Arguments

...

Rest arguments.

Method clone()

The objects of this class are cloneable with this method.

Usage

PIOutputMapping$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.