PIOutputMapping

Establishes connections between simulated quantities and corresponding observed data sets. Utilized within ParameterIdentification instances to align and compare simulation outputs with empirical data.

Super class

ospsuite.utils::Printable -> PIOutputMapping

Active bindings

observedDataSets

A named list containing DataSet objects for comparison with simulation outcomes.

dataTransformations

A named list of factors and offsets.

quantity

Simulation quantities to be aligned with observed data values.

scaling

Specifies scaling for output mapping: linear (default) or logarithmic.

transformResultsFunction

A function to preprocess simulated results (time and observation values) before residual calculation. It takes numeric vectors 'xVals' and 'yVals', and returns a named list with keys 'xVals' and 'yVals'.

Methods

Public methods

PIOutputMapping$new()

PIOutputMapping$addObservedDataSets()

PIOutputMapping$removeObservedDataSet()

PIOutputMapping$setDataTransformations()

PIOutputMapping$print()

PIOutputMapping$clone()

Method new()

Initialize a new instance of the class.

Usage

PIOutputMapping$new(quantity)

Arguments

quantity

An object of the type Quantity.

Returns

A new PIOutputMapping object. Adds or updates observed data using DataSet objects.

Method addObservedDataSets()

Usage

PIOutputMapping$addObservedDataSets(data)

Arguments

data

A DataSet object or a list thereof, matching the simulation quantity dimensions.

Details

Replaces any existing data set with the same label.

Method removeObservedDataSet()

Removes specified observed data series.

Usage

PIOutputMapping$removeObservedDataSet(label)

Arguments

label

The label of the observed data series to remove.

Method setDataTransformations()

Configures transformations for dataset(s).

Usage

PIOutputMapping$setDataTransformations(
  labels = NULL,
  xOffsets = 0,
  yOffsets = 0,
  xFactors = 1,
  yFactors = 1
)

Arguments

labels

List of dataset labels for targeted transformations. Absence of labels applies transformations globally.

xOffsets

Numeric list/value for X-offset adjustments.

yOffsets

Numeric list/value for Y-offset adjustments.

xFactors

Numeric list/value for X-scaling factors.

yFactors

Numeric list/value for Y-scaling factors.

Method print()

Prints a summary of the PIOutputMapping.

Usage

PIOutputMapping$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

PIOutputMapping$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.