ParameterIdentification

Executes optimization to find the best parameter values by comparing simulation results against observed data.

Value

A new ParameterIdentification object. Executes Parameter Identification

Super class

ospsuite.utils::Printable -> ParameterIdentification

Active bindings

simulations

A named list of simulation objects, keyed by the IDs of their root containers.

parameters

Read-only list of PIParameters objects for optimization.

configuration

A PIConfiguration object instance.

outputMappings

A list of PIOutputMapping objects linking observed data to simulation outputs.

Methods

Public methods

ParameterIdentification$new()

ParameterIdentification$run()

ParameterIdentification$plotResults()

ParameterIdentification$print()

Method new()

Initializes a ParameterIdentification instance.

Usage

ParameterIdentification$new(
  simulations,
  parameters,
  outputMappings,
  configuration = NULL
)

Arguments

simulations

An object or a list of objects of class Simulation. Parameters of the simulation object will be varied and the results simulated. For creating Simulation objects, see ospsuite::loadSimulation.

parameters

An object or a list of objects of class PIParameter. These parameters will be varied. For creating PIParameter objects, refer to PIParameters.

outputMappings

List of objects of the class PIOutputMapping. Each object maps a model output (represented by a Quantity) with a set of observed data given as XYData objects. For guidance on creating PIOutputMapping objects, see PIOutputMapping.

configuration

(Optional) PIConfiguration for additional settings. Uses default if omitted. For details on creating a PIConfiguration object, see PIConfiguration.

Method run()

Initiates parameter identification process. Upon completion, access optimal parameter values through PIParameters$currValue.

Usage

ParameterIdentification$run()

Returns

Results from the parameter identification algorithm, varying by algorithm choice. Plots Parameter Estimation Results

Method plotResults()

Generates plots for each output mapping based on the current or provided parameter values. Simulations are executed with these parameters to visualize the estimation results.

Usage

ParameterIdentification$plotResults(par = NULL)

Arguments

par

Optional parameter values for simulations, in the order of ParameterIdentification$parameters. Use current values if NULL.

Returns

A list of ggplot2 plots, one for each PIOutputMapping.

Method print()

Prints a summary of ParameterIdentification instance.

Usage

ParameterIdentification$print()