ParameterIdentification

A task to identify optimal parameter values based on simulation outputs and observed data

Super class

ospsuite.utils::Printable -> ParameterIdentification

Active bindings

simulations

Named list with simulation objects, where names are IDs of the root container of the simulation

parameters

List of PIParameters objects to be optimized. Read-only

configuration

An object of PIConfiguration

outputMappings

List of PIOutputMapping objects. Each mapping assigns a set of observed data to a simulation output Evaluate all simulations with given parameter values

Methods

Public methods

ParameterIdentification$new()

ParameterIdentification$run()

ParameterIdentification$plotResults()

ParameterIdentification$gridSearch()

ParameterIdentification$calculateOFVProfiles()

ParameterIdentification$plotOFVProfiles()

ParameterIdentification$plotGrid()

ParameterIdentification$print()

Method new()

Initialize a new instance of the class

Usage

ParameterIdentification$new(
  simulations,
  parameters,
  outputMappings,
  configuration = NULL
)

Arguments

simulations

An object or a list of objects of class Simulation. Parameters of the simulation object will be varied and the results simulated

parameters

An object or a list of objects of class PIParameter. These parameters will be varied.

outputMappings

List of objects of the class PIOutputMapping. Each objects maps a model output (represented by a Quantity) with a set of observed data given as XYData objects. is used.

configuration

Optional. Object of type PIConfiguration defining further options of the parameter identification. If no PIConfiguration is passed, a default one

currVals

Numerical vector of the parameter values to be applied

Returns

An list of objects of DataCombined class that includes values simulated with the given parameters, and corresponding datasets with observed data. Returns one DataCombined object for each output mapping.

Method run()

Start identification of parameters

Usage

ParameterIdentification$run()

Details

When the identification if finished, the best identified values of the parameters are accessible via the currValue-field of the PIParameters-object.

Returns

Output of the PI algorithm. Depends on the selected algorithm.

Method plotResults()

Plot the results of parameter estimation

Usage

ParameterIdentification$plotResults(par = NULL)

Arguments

par

Values of paramterers to be applied to the simulations. If NULL (default), current parameter values are applied. If custom values are supplied, the they must be in the same order as ParameterIdentification$parameters

Details

Runs all simulations with current (default) or supplied parameter values and creates plots of every output mapping

Method gridSearch()

Calculates the values of the objective function on a rectangular grid, and optionally saves the best result as the starting point for next optimization runs.

Usage

ParameterIdentification$gridSearch(
  lower = NA,
  upper = NA,
  logScaleFlag = FALSE,
  totalEvaluations = 50,
  margin = 0,
  setStartingPoint = FALSE
)

Arguments

lower

A vector of lower bounds for parameters, with the same length as the number of parameters. By default, uses the minimal values supported for parameters.

upper

A vector of upper bounds for parameters, with the same length as the number of parameters. By default, uses the maximal values supported for parameters.

logScaleFlag

A single logical value or a vector of logical values indicating if grid should be evenly spaced on a linear or a logarithmic scale. Defaults to FALSE.

totalEvaluations

An integer number. The grid will have as many points so that the total number of grid points does not exceed totalEvaluations. Defaults to 50.

margin

Can be set to a non-zero positive value so that the edges of the grid will be away from the exact parameter bounds.

setStartingPoint

(logical) If TRUE, the best result will be saved as the starting point for the next optimization runs

Returns

A tibble with one column for each parameter and one column for the objective function value. The tibble will have at most totalEvaluations rows.

Method calculateOFVProfiles()

Calculates the values of the objective function on all orthogonal lines passing through a given point in the parameter space.

Usage

ParameterIdentification$calculateOFVProfiles(
  par = NA,
  lower = NA,
  upper = NA,
  totalEvaluations = NA
)

Arguments

par

A vector of parameter values, with the same length as the number of parameters. If not supplied, the current parameter values are used.

lower

A vector of lower bounds for parameters, with the same length as the number of parameters. By default, uses 0.9 of the current parameter value.

upper

A vector of upper bounds for parameters, with the same length as the number of parameters. By default, uses 1.1 of the current parameter value.

totalEvaluations

An integer number. The combined profiles will not contain more than totalEvaluations points. If not supplied, 21 points per parameter are plotted to cover a uniform grid from 0.9 to 1.1.

Returns

A list of tibbles, one tibble per parameter, with one column for parameter values and one column for the matching objective function values.

Method plotOFVProfiles()

Plot the profiles of the objective function calculated by the calculateOFVProfiles() method.

Usage

ParameterIdentification$plotOFVProfiles(profiles)

Arguments

profiles

A list of tibbles, with each tibble containing two columns. The first column should contain several values for the parameter, and the second column, called ofv, should contain the matching objective function value.

Returns

A list of ggplot2 objects, one for each parameter.

Method plotGrid()

Plot a heatmap of the objective function calculated by the gridSearch() method

Usage

ParameterIdentification$plotGrid(grid)

Arguments

grid

A tibble with three columns. The first two columns should contain parameter values for a 2-parameter optimization problem, and the ofv column should contain the matching objective function values.

Returns

A ggplot2 object.

Method print()

Print the object to the console

Usage

ParameterIdentification$print(...)

Arguments

...

Rest arguments.